(MOT-4664) feat(voice): accurate two-pass dictation, console page redesign, configuration form
Dictation now runs two recognizers: the streaming Zipformer keeps showing
words as you speak and decides where a sentence ends, and Parakeet TDT
0.6B v2 re-decodes each finished sentence so the committed text has
punctuation, casing and far fewer misses. The accurate model downloads in
the background on first use and the streaming text is used until it is
ready. A faint preroll and endpoint reset only on non-empty text stop the
streaming model from dropping the first words of a sentence.

The console page is rebuilt on the shared page shell: Overview, Dictate,
Transcribe, Models and Read aloud sections with a narrow-width drawer,
container queries and touch rules. Every choice is explicit and written to
the worker configuration from the page itself: which model transcribes,
which one shows live words, which read-aloud engine runs. The Models
section lists what each model is for, whether it is on disk, and offers
download, use and remove. A configuration form registered through
host.configForms puts every WorkerConfig field into global Settings in
plain language.

Adds voice::models::remove, a kind field on catalog entries, final-model
state in voice::doctor, and stores endpoint timings as f64 so the
configuration round-trips the values a person typed.

+++ voice/src/config.rs
@@ -86,18 +86,18 @@ pub struct SttConfig {
     /// Trailing silence, in seconds, that ends an utterance once speech has
     /// been heard. Lower commits dictated text sooner; higher tolerates pauses.
     #[serde(default = "default_silence_after_speech_secs")]
-    pub silence_after_speech_secs: f32,
+    pub silence_after_speech_secs: f64,
 
     /// Trailing silence, in seconds, that ends a segment when nothing has been
     /// recognized yet. Keep this well above the model's lookahead so the
     /// first words of an utterance are never cut.
     #[serde(default = "default_silence_without_speech_secs")]
-    pub silence_without_speech_secs: f32,
+    pub silence_without_speech_secs: f64,
 
     /// Longest single utterance, in seconds, before the recognizer commits
     /// what it has and starts a new segment.
     #[serde(default = "default_max_utterance_secs")]
-    pub max_utterance_secs: f32,
+    pub max_utterance_secs: f64,
 
     /// Settings for the `openai` backend.
     #[serde(default)]
@@ -219,15 +219,15 @@ fn default_num_threads() -> usize {
     2
 }
 
-fn default_silence_after_speech_secs() -> f32 {
+fn default_silence_after_speech_secs() -> f64 {
     0.8
 }
 
-fn default_silence_without_speech_secs() -> f32 {
+fn default_silence_without_speech_secs() -> f64 {
     2.4
 }
 
-fn default_max_utterance_secs() -> f32 {
+fn default_max_utterance_secs() -> f64 {
     20.0
 }
